Only delete dest file before copy if it exists

+++ app/src/main/java/ca/josephroque/bowlingcompanion/utils/Files.kt
@@ -38,7 +38,9 @@ object Files {
     fun copyFile(source: File, dest: File): Deferred<Boolean> {
         val bufferSize = 1024
         return async(CommonPool) {
-            dest.delete()
+            if (dest.exists()) {
+                dest.delete()
+            }
 
             var inputStream: InputStream? = null
             var outputStream: OutputStream? = null
